lsass: 64-bit fixes.
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@55478 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lsass/client/local/privileges.c
@@ -186,9 +186,9 @@ LsaLocalLookupPrivilegeValue(
 {
     DWORD dwError = ERROR_SUCCESS;
     size_t privilegeNameLen = 0;
-    size_t requestSize = 0;
+    DWORD requestSize = 0;
     PVOID pRequestBlob = NULL;
-    size_t responseSize = 0;
+    DWORD responseSize = 0;
     PVOID pResponseBlob = NULL;
 
     dwError = LwWc16sLen(pwszPrivilegeName,
@@ -237,9 +237,9 @@ LsaLocalLookupPrivilegeName(
     )
 {
     DWORD dwError = ERROR_SUCCESS;
-    size_t requestSize = 0;
+    DWORD requestSize = 0;
     PVOID pRequestBlob = NULL;
-    size_t responseSize = 0;
+    DWORD responseSize = 0;
     PVOID pResponseBlob = NULL;
 
     requestSize  = sizeof(*pPrivilegeValue);
@@ -280,9 +280,9 @@ LsaLocalEnumAccountRights(
     LWMsgDataContext* pDataContext = NULL;
     PSTR pszAccountSid = NULL;
     PLSA_LOCAL_IPC_ENUM_ACCOUNT_RIGHTS_RESP pResponse = NULL;
-    size_t requestSize = 0;
+    DWORD requestSize = 0;
     PVOID pRequestBlob = NULL;
-    size_t responseSize = 0;
+    DWORD responseSize = 0;
     PVOID pResponseBlob = NULL;
 
     dwError = MAP_LWMSG_ERROR(lwmsg_context_new(NULL, &context));
